Refactor markdown utils
